Last commit before moving to Perlmutter/Dante and crying
RIP Cori

+++ libmidapack/include/spherical_harmonics/s2hat_tools.h
@@ -138,7 +138,7 @@ int apply_alm2pix(double *local_map_pix, s2hat_dcomplex *local_alm, S2HAT_parame
 int apply_pix2alm(double *local_map_pix, s2hat_dcomplex *local_alm, S2HAT_parameters *S2HAT_params);
 /**/
 /* Apply inverse of covariance matrix to local_alm */
-int apply_inv_covariance_matrix_to_alm(s2hat_dcomplex *input_local_alm, s2hat_dcomplex *out_local_alm, double **inv_covariance_matrix, int power_inv_cov, S2HAT_parameters *S2HAT_params);/**/
+int apply_inv_covariance_matrix_to_alm(s2hat_dcomplex *input_local_alm, s2hat_dcomplex *out_local_alm, double **inv_covariance_matrix, double power_inv_cov, S2HAT_parameters *S2HAT_params);/**/
 /* Transform alm to c_ell coefficients */
 int alm2cls(s2hat_dcomplex* local_alm, double *c_ell_array, int nspec, S2HAT_parameters *S2HAT_params);
 /**/
